[sw/lib] delay function no longer needs CYCLE CSRs
instead an ASM loop is used

+++ sw/lib/include/neorv32_cpu.h
@@ -50,7 +50,7 @@ void neorv32_cpu_set_mcycle(uint64_t value);
 uint64_t neorv32_cpu_get_instret(void);
 void neorv32_cpu_set_minstret(uint64_t value);
 uint64_t neorv32_cpu_get_systime(void);
-void neorv32_cpu_delay_ms(uint32_t time_ms);
+void neorv32_cpu_delay_ms(int16_t time_ms);
 void __attribute__((naked)) neorv32_cpu_goto_user_mode(void);
 int neorv32_cpu_atomic_cas(uint32_t addr, uint32_t expected, uint32_t desired);
 uint32_t neorv32_cpu_pmp_get_num_regions(void);
